+++ query/src/lib/base-query.ts
@@ -12,6 +12,7 @@ import {
 } from '@tanstack/query-core';
 import { Observable, shareReplay } from 'rxjs';
 import { normalizeOptions } from './query-options';
+import { shouldThrowError } from './utils';
 
 export type QueryFunctionWithObservable<
   T = unknown,
@@ -25,7 +26,7 @@ export interface Options {
   injector?: Injector;
 }
 
-interface _CreateBaseQueryOptions<
+export interface _CreateBaseQueryOptions<
   TQueryFnData = unknown,
   TError = DefaultError,
   TData = TQueryFnData,
@@ -121,11 +122,22 @@ export function createBaseQuery<
 
     const queryObserverDispose = queryObserver.subscribe(
       notifyManager.batchCalls((result) => {
-        observer.next(
-          defaultedOptions.notifyOnChangeProps
-            ? result
-            : queryObserver?.trackResult(result),
-        );
+        if (
+          !result.isFetching &&
+          result.isError &&
+          shouldThrowError(queryObserver!.options.throwOnError, [
+            result.error,
+            queryObserver!.getCurrentQuery(),
+          ])
+        ) {
+          observer.error(result.error);
+        } else {
+          observer.next(
+            defaultedOptions.notifyOnChangeProps
+              ? result
+              : queryObserver?.trackResult(result),
+          );
+        }
       }),
     );